% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wishart.R
\name{rwishart_chol}
\alias{rwishart_chol}
\title{Sampling Cholesky factor of a Wishart matrix}
\usage{
rwishart_chol(n, nu, Sigma)
}
\arguments{
\item{n}{sample size, a positive integer}

\item{nu}{degrees of freedom, a positive number at least equal to the dimension
(the order of \code{Sigma})}

\item{Sigma}{scale matrix, a positive definite real matrix}
}
\value{
A numeric three-dimensional array;
simulations are stacked along the third dimension (see example).
}
\description{
Samples the lower triangular Cholesky factor of a Wishart random matrix.
}
\examples{
nu <- 4
p <- 3
Sigma <- diag(p)
Wsims <- rwishart_chol(10000, nu, Sigma)
dim(Wsims) # 3 3 10000
Wsims[,,1]
}
