% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta.R
\name{rmatrixbeta}
\alias{rmatrixbeta}
\title{Matrix Beta sampler}
\usage{
rmatrixbeta(n, p, a, b, Theta1 = NULL, Theta2 = NULL)
}
\arguments{
\item{n}{sample size, a positive integer}

\item{p}{dimension, a positive integer}

\item{a, b}{parameters of the distribution, positive numbers with constraints given
in Details}

\item{Theta1}{numerator noncentrality parameter, a positive semidefinite real
matrix of order \code{p}; setting it to \code{NULL} (default) is
equivalent to setting it to the zero matrix}

\item{Theta2}{denominator noncentrality parameter, a positive semidefinite real
matrix of order \code{p}; setting it to \code{NULL} (default) is
equivalent to setting it to the zero matrix}
}
\value{
A numeric three-dimensional array;
simulations are stacked along the third dimension (see example).
}
\description{
Samples a matrix Beta distribution.
}
\details{
Parameters \code{a} and \code{b} are positive numbers that satisfy the
following constraints:
\itemize{
\item if both \code{Theta1} and \code{Theta2} are the null matrix,
\code{a+b >= p/2}; if \code{a < p/2}, it must be half an integer;
if \code{b < p/2}, it must be half an integer
\item if \code{Theta1} is not the null matrix, \code{a >= p/2};
if \code{b < p/2}, it must be half an integer
\item if \code{Theta2} is the null matrix, \code{b >= p/2};
if \code{a < p/2}, it must be half an integer}
}
\examples{
Bsims <- rmatrixbeta(10000, 3, 1, 1)
dim(Bsims) # 3 3 10000
}
