% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta.R
\name{rmatrixbetaII}
\alias{rmatrixbetaII}
\title{Matrix Beta II sampler}
\usage{
rmatrixbetaII(n, p, a, b, Theta1 = NULL, Theta2 = NULL)
}
\arguments{
\item{n}{sample size, a positive integer}

\item{p}{dimension, a positive integer}

\item{a, b}{parameters of the distribution, positive numbers with constraints given
in Details}

\item{Theta1}{numerator noncentrality parameter, a positive semidefinite real
matrix of order \code{p}; setting it to \code{NULL} (default) is
equivalent to setting it to the zero matrix}

\item{Theta2}{denominator noncentrality parameter, a positive semidefinite real
matrix of order \code{p}; setting it to \code{NULL} (default) is
equivalent to setting it to the zero matrix}
}
\value{
A numeric three-dimensional array;
simulations are stacked along the third dimension (see example).
}
\description{
Samples a matrix Beta type II distribution.
}
\details{
Parameters \code{a} and \code{b} are positive numbers that satisfy the
following constraints:
\itemize{
\item in any case, \code{b >= p/2}
\item if \code{Theta1} is not the null matrix, \code{a >= p/2}
\item if \code{Theta1} is the null matrix and \code{a < p/2}, then \code{a}
must be half an integer}
}
\examples{
Bsims <- rmatrixbetaII(10000, 3, 1, 1)
dim(Bsims) # 3 3 10000
}
