% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wishart.R
\name{rwishart}
\alias{rwishart}
\title{Wishart sampler}
\usage{
rwishart(n, nu, Sigma, Theta = NULL)
}
\arguments{
\item{n}{sample size, a positive integer}

\item{nu}{degrees of freedom, a positive number;
if less than the dimension (the order of \code{Sigma}), must be an integer;
in the noncentral case (i.e. when \code{Theta} is not the null matrix),
\itemize{
\item \code{nu} must be at least equal to the dimension \code{p}
\item if \code{nu<2*p}, it must be an integer}}

\item{Sigma}{scale matrix, a positive semidefinite real matrix}

\item{Theta}{noncentrality parameter, a positive semidefinite real matrix of
same dimension as \code{Sigma}; setting it to \code{NULL} (default) is
equivalent to setting it to the zero matrix}
}
\value{
A numeric three-dimensional array;
simulations are stacked along the third dimension (see example).
}
\description{
Samples a Wishart distribution.
}
\details{
A sampled Wishart matrix is always positive semidefinite. If \code{nu}
is at least equal to the dimension (the order of \code{Sigma}), it is positive
definite.
}
\examples{
nu <- 4
p <- 3
Sigma <- toeplitz(p:1)
Theta <- diag(p)
Wsims <- rwishart(10000, nu, Sigma, Theta)
dim(Wsims) # 3 3 10000
apply(Wsims, 1:2, mean) # approximately nu*Sigma+Theta
}
